+++ deployment/helm/posix-mapper/CHANGELOG.md
@@ -1,4 +1,8 @@
-# CHANGELOG for POSIX Mapper (Chart 0.1.8)
+# CHANGELOG for POSIX Mapper (Chart 0.2.0)
+
+## 2024.04.18 (0.2.0)
+- Configure authorized groups that are allowed to access POSIX Mapper data
+- Configure Kubernetes secret name for trusted service to service calls
 
 ## 2023.11.02 (0.1.8)
 - Swagger documentation fix (Bug)